메뉴들 불러오기, 리뷰 작성 중

+++ FaMin copy/app/src/main/java/com/example/famin/fragment/MarketInfo/MenuListModel.kt
@@ -3,6 +3,6 @@ package com.example.famin.fragment.MarketInfo
 data class MenuListModel(
     var image : String,
     var name : String,
-    var price : Int,
+    var price : Long,
     var info : String
 )
